fix(nutrition): make meal-entry creation atomic with transactional writes (#147)
Extract MealEntryWriteService to hold @Transactional create/update/delete
operations, invoked synchronously from MealEntryService's boundedElastic
callables so Spring's transactional proxy applies correctly. Make
DayTargetSnapshotService.ensureSnapshot idempotent under concurrency by
running it in its own REQUIRES_NEW transaction and using saveAndFlush so a
unique-constraint violation from a concurrent first-entry is swallowed
without affecting the meal-entry transaction.
