+++ nutrition/src/main/java/com/marvin/nutrition/service/MealEntryService.java
@@ -14,7 +14,6 @@ import com.marvin.nutrition.repository.DayTargetSnapshotRepository;
 import com.marvin.nutrition.repository.FoodRepository;
 import com.marvin.nutrition.repository.MealEntryRepository;
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -30,14 +29,12 @@ import reactor.core.scheduler.Schedulers;
 @Service
 public class MealEntryService {
 
-    private static final BigDecimal HUNDRED = BigDecimal.valueOf(100);
-
     private final MealEntryRepository mealEntryRepository;
     private final FoodRepository foodRepository;
     private final MealEntryMapper mealEntryMapper;
     private final NutritionTargetService nutritionTargetService;
     private final DayTargetSnapshotRepository dayTargetSnapshotRepository;
-    private final DayTargetSnapshotService dayTargetSnapshotService;
+    private final MealEntryWriteService mealEntryWriteService;
 
     /**
      * Creates a new MealEntryService with the required dependencies.
@@ -47,7 +44,7 @@ public class MealEntryService {
      * @param mealEntryMapper             MapStruct mapper for entity/DTO conversion
      * @param nutritionTargetService      service for computing daily nutrition targets
      * @param dayTargetSnapshotRepository JPA repository for per-day nutrition target snapshots
-     * @param dayTargetSnapshotService    service for creating/refreshing per-day nutrition target snapshots
+     * @param mealEntryWriteService       service owning transactional meal entry write operations
      */
     public MealEntryService(
             MealEntryRepository mealEntryRepository,
@@ -55,13 +52,13 @@ public class MealEntryService {
             MealEntryMapper mealEntryMapper,
             NutritionTargetService nutritionTargetService,
             DayTargetSnapshotRepository dayTargetSnapshotRepository,
-            DayTargetSnapshotService dayTargetSnapshotService) {
+            MealEntryWriteService mealEntryWriteService) {
         this.mealEntryRepository = mealEntryRepository;
         this.foodRepository = foodRepository;
         this.mealEntryMapper = mealEntryMapper;
         this.nutritionTargetService = nutritionTargetService;
         this.dayTargetSnapshotRepository = dayTargetSnapshotRepository;
-        this.dayTargetSnapshotService = dayTargetSnapshotService;
+        this.mealEntryWriteService = mealEntryWriteService;
     }
 
     /**
@@ -77,23 +74,8 @@ public class MealEntryService {
      * @return a Mono emitting the created meal entry DTO
      */
     public Mono<MealEntryDTO> addEntry(LocalDate date, CreateMealEntryRequest req) {
-        return Mono.fromCallable(() -> {
-            final MealEntryEntity entity = new MealEntryEntity();
-            entity.setEntryDate(date);
-            entity.setMealType(req.mealType());
-
-            final MealEntryDTO dto;
-            if (req.foodId() != null) {
-                final String foodName = buildFoodEntry(entity, req);
-                dto = mealEntryMapper.toDTO(mealEntryRepository.save(entity), foodName);
-            } else {
-                buildAdHocEntry(entity, req);
-                dto = mealEntryMapper.toDTO(mealEntryRepository.save(entity));
-            }
-
-            dayTargetSnapshotService.ensureSnapshot(date);
-            return dto;
-        }).subscribeOn(Schedulers.boundedElastic());
+        return Mono.fromCallable(() -> mealEntryWriteService.createEntry(date, req))
+                .subscribeOn(Schedulers.boundedElastic());
     }
 
     /**
@@ -108,22 +90,8 @@ public class MealEntryService {
      * @return a Mono emitting the updated meal entry DTO
      */
     public Mono<MealEntryDTO> updateEntry(UUID id, UpdateMealEntryRequest req) {
-        return Mono.fromCallable(() -> {
-            final MealEntryEntity entity = mealEntryRepository.findById(id)
-                    .orElseThrow(() -> new NoSuchElementException("Meal entry not found: " + id));
-
-            if (req.mealType() != null) {
-                entity.setMealType(req.mealType());
-            }
-
-            if (entity.getFoodId() != null) {
-                final String foodName = applyFoodEntryUpdate(entity, req);
-                return mealEntryMapper.toDTO(mealEntryRepository.save(entity), foodName);
-            }
-
-            applyAdHocEntryUpdate(entity, req);
-            return mealEntryMapper.toDTO(mealEntryRepository.save(entity));
-        }).subscribeOn(Schedulers.boundedElastic());
+        return Mono.fromCallable(() -> mealEntryWriteService.updateEntry(id, req))
+                .subscribeOn(Schedulers.boundedElastic());
     }
 
     /**
@@ -135,9 +103,7 @@ public class MealEntryService {
      */
     public Mono<Void> deleteEntry(UUID id) {
         return Mono.fromCallable(() -> {
-            final MealEntryEntity entity = mealEntryRepository.findById(id)
-                    .orElseThrow(() -> new NoSuchElementException("Meal entry not found: " + id));
-            mealEntryRepository.delete(entity);
+            mealEntryWriteService.deleteEntry(id);
             return null;
         }).subscribeOn(Schedulers.boundedElastic()).then();
     }
@@ -203,111 +169,6 @@ public class MealEntryService {
                 snapshot.getCarbsG(),
                 snapshot.getBasis()
         );
-    }
-
-    /**
-     * Populates the entity fields for a food-backed entry by looking up the food and snapshotting macros.
-     *
-     * @param entity the entity to populate
-     * @param req    the create request
-     * @return the name of the referenced food item
-     */
-    private String buildFoodEntry(MealEntryEntity entity, CreateMealEntryRequest req) {
-        final FoodEntity food = foodRepository.findById(req.foodId())
-                .orElseThrow(() -> new NoSuchElementException("Food not found: " + req.foodId()));
-
-        if (req.quantityG() == null) {
-            throw new IllegalArgumentException("quantityG is required for a food entry");
-        }
-
-        entity.setFoodId(req.foodId());
-        entity.setQuantityG(req.quantityG());
-        entity.setDescription(req.description());
-        entity.setKcal(snapshot(food.getKcalPer100(), req.quantityG()));
-        entity.setProteinG(snapshot(food.getProteinPer100(), req.quantityG()));
-        entity.setCarbsG(snapshot(food.getCarbsPer100(), req.quantityG()));
-        entity.setFatG(snapshot(food.getFatPer100(), req.quantityG()));
-        entity.setFoodName(food.getName());
-        return food.getName();
-    }
-
-    /**
-     * Populates the entity fields for an ad-hoc entry from the request.
-     *
-     * @param entity the entity to populate
-     * @param req    the create request
-     */
-    private void buildAdHocEntry(MealEntryEntity entity, CreateMealEntryRequest req) {
-        if (req.description() == null || req.description().isBlank()
-                || req.kcal() == null || req.proteinG() == null
-                || req.carbsG() == null || req.fatG() == null) {
-            throw new IllegalArgumentException(
-                    "ad-hoc entry requires description and kcal/proteinG/carbsG/fatG");
-        }
-        entity.setDescription(req.description());
-        entity.setKcal(req.kcal());
-        entity.setProteinG(req.proteinG());
-        entity.setCarbsG(req.carbsG());
-        entity.setFatG(req.fatG());
-    }
-
-    /**
-     * Applies update fields for a food-backed entry, re-snapshotting macros if quantity changed.
-     * Always resolves and returns the food name so the caller can populate {@code foodName} on the DTO.
-     *
-     * @param entity the entity to update
-     * @param req    the update request
-     * @return the name of the referenced food item
-     */
-    private String applyFoodEntryUpdate(MealEntryEntity entity, UpdateMealEntryRequest req) {
-        final FoodEntity food = foodRepository.findById(entity.getFoodId())
-                .orElseThrow(() -> new NoSuchElementException("Food not found: " + entity.getFoodId()));
-        if (req.quantityG() != null) {
-            entity.setQuantityG(req.quantityG());
-            entity.setKcal(snapshot(food.getKcalPer100(), req.quantityG()));
-            entity.setProteinG(snapshot(food.getProteinPer100(), req.quantityG()));
-            entity.setCarbsG(snapshot(food.getCarbsPer100(), req.quantityG()));
-            entity.setFatG(snapshot(food.getFatPer100(), req.quantityG()));
-        }
-        entity.setFoodName(food.getName());
-        return food.getName();
-    }
-
-    /**
-     * Applies update fields for an ad-hoc entry.
-     * Note: {@code req.quantityG()} is intentionally ignored here — it is only meaningful for
-     * food-backed entries where it drives macro re-snapshotting from the food catalog.
-     *
-     * @param entity the entity to update
-     * @param req    the update request
-     */
-    private void applyAdHocEntryUpdate(MealEntryEntity entity, UpdateMealEntryRequest req) {
-        if (req.description() != null) {
-            entity.setDescription(req.description());
-        }
-        if (req.kcal() != null) {
-            entity.setKcal(req.kcal());
-        }
-        if (req.proteinG() != null) {
-            entity.setProteinG(req.proteinG());
-        }
-        if (req.carbsG() != null) {
-            entity.setCarbsG(req.carbsG());
-        }
-        if (req.fatG() != null) {
-            entity.setFatG(req.fatG());
-        }
-    }
-
-    /**
-     * Computes the snapshotted macro value: {@code per100 × quantityG / 100}, rounded half-up to 2 decimals.
-     *
-     * @param per100    the per-100-g value from the food catalog
-     * @param quantityG the portion size in grams
-     * @return the snapshotted value
-     */
-    private BigDecimal snapshot(BigDecimal per100, BigDecimal quantityG) {
-        return per100.multiply(quantityG).divide(HUNDRED, 2, RoundingMode.HALF_UP);
     }
 
     /**
